Pathfinding appears to be working, though does not terminate when needed
previously added the neighbors of every node in the open set. fixed it to only add nodes that are neighbors of the current node. Based on priliminary testing, it will get to the right node in an expected path, but will then continue to search the map and add nodes to the map