Add dbreader.js
This file is used to access the database and get information. At the moment you may run getRoom(), which will get info from a textbox called 'teacher' and a text box called 'period'. This can be changed later. getRoom will get the infor and run the function getInfo(), which will get the information from the XML file. getRoom() may be edited so that it uses the proper HTML elements, but do not touch the getInfo() function please. I will also be adding functionality for first names and will be returning a list of rooms in case there are duplicate entries(in the case of shared last names)