File changes image to text
The python file is ran, then changes the pixels to a text file. The first two lines are the width and height of the image, then the next line if all the pixels, so that each pixel color is separated by a semi-colon and each pixel color is represented by its RGB value, where each number is separated by a comma. It looks like this

width
height
r1,g1,b1;r2g,2b,2;r3g,3b,3;r4g,4b,4r;5g,5b,5r;6g,6b,6r;   and so on for all pixels in the image